requirements.txt
discord.py>=2.4.0
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.0
Pillow>=10.0.0